Refactor XNNPACK tester to extract base classes (#11596)
Summary:
Refactor the XNNPACK tester to split out reusable base components from XNNPACK-specific parts. I've relocated the base classes to backends/test/harness.

I've kept the tester structure pretty much unchanged, except for replacing stage names with an enum.

It looks like Arm tests are directly importing for XNNPACK's tester currently. Ideally, we'll want to refactor to have their own stage implementations, but I've left that as a follow-up to minimize changes for the initial refactor.

Pull Request resolved: https://github.com/pytorch/executorch/pull/11596

Test Plan:
CI

Rollback Plan:

Differential Revision: D76547310

Pulled By: GregoryComer
