+++ backends/test/harness/__init__.py
@@ -1,0 +1,3 @@
+from .tester import Tester
+
+__all__ = ["Tester"]
